move git_root outside of codefilemanager and move git root tests to new file
